inductor_config_logging: Don't drop keys
This bit me while I was trying to debug some trace issues.
In general this config is already quite large when dumping, so adding
more fields doesn't make it significantly worse.

Also a number of the items we are type checking for (except the test
configs), don't even show up. Primarily this will help us when debugging
rocm, halide, and trace configs.

ghstack-source-id: f17ca2d531d009e2de11b16bdba6816947f19811
Pull Request resolved: https://github.com/pytorch/pytorch/pull/144700